feat: enable multiple trees
