+++ package.json
@@ -17,38 +17,32 @@
 		"singleQuote": true
   },
   "devDependencies": {
-    "@types/react": "^18.0.20",
-    "@types/styled-components": "^5.1.26",
-    "@typescript-eslint/eslint-plugin": "^5.0.0",
-    "@typescript-eslint/parser": "^5.38.0",
-    "contentful-management": "^10.6.3",
-    "contentful-ui-extensions-sdk": "^4.8.1",
-    "eslint": "^8.0.1",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-config-standard-with-typescript": "^23.0.0",
-    "eslint-plugin-import": "^2.25.2",
-    "eslint-plugin-json": "3.1.0",
-    "eslint-plugin-n": "^15.0.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-promise": "^6.0.0",
-    "eslint-plugin-react": "^7.31.8",
-    "immer": "^9.0.15",
-    "install-peers": "^1.0.4",
-    "prettier": "2.7.1",
-    "react": "18.2.0",
-    "styled-components": "^5.3.5",
-    "typescript": "*",
-    "use-immer": "^0.7.0"
-  },
-  "peerDependencies": {
-    "install-peers": "^1.0.4"
-  },
-  "dependencies": {
-    "contentful-management": "^10.15.0",
-    "contentful-ui-extensions-sdk": "^4.12.1",
-    "immer": "^9.0.15",
-    "react": "^18.2.0",
-    "styled-components": "^5.3.5",
-    "use-immer": "^0.7.0"
-  }
+		"@types/react": "^18.0.20",
+		"@types/styled-components": "^5.1.26",
+		"@typescript-eslint/eslint-plugin": "^5.0.0",
+		"@typescript-eslint/parser": "^5.38.0",
+		"eslint": "^8.0.1",
+		"eslint-config-prettier": "^8.5.0",
+		"eslint-config-standard-with-typescript": "^23.0.0",
+		"eslint-plugin-import": "^2.25.2",
+		"eslint-plugin-json": "3.1.0",
+		"eslint-plugin-n": "^15.0.0",
+		"eslint-plugin-prettier": "^4.2.1",
+		"eslint-plugin-promise": "^6.0.0",
+		"eslint-plugin-react": "^7.31.8",
+		"install-peers": "^1.0.4",
+		"prettier": "2.7.1",
+		"typescript": "*"
+	},
+	"peerDependencies": {
+		"install-peers": "^1.0.4"
+	},
+	"dependencies": {
+		"contentful-management": "10.15.0",
+		"contentful-ui-extensions-sdk": "4.12.1",
+		"immer": "9.0.15",
+		"react": "18.2.0",
+		"styled-components": "5.3.5",
+		"use-immer": "0.7.0"
+	}
 }
